desafio/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.6.1
cryptography==43.0.1
Django==5.0.8
django-allauth==0.63.6
django-allauth-2fa==0.11.1
django-otp==1.2.3
idna==3.7
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pypng==0.20220715.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
reportlab==4.2.2
django-simple-captcha==0.6.0
redis==5.1.0 
celery==5.4.0